pyproject.toml
[project]
name = "codarme"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "guilherme", email = "vinespedro55@gmail.com"},
]
dependencies = [
    "pylint>=3.0.3",
    "flake8>=7.0.0",
    "django>=5.0.4",
    "djangorestframework>=3.15.1",
    "pytest-django>=4.8.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.1",
]
requires-python = ">=3.11"
readme = "README.md"
license = {text = "MIT"}

[tool.pdm]
distribution = true

[tool.pdm.dev-dependencies]
dev = [
    "pylint>=3.0.3",
    "autopep8>=2.0.4",
    "pytest>=8.0.1",
]

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"
